Remove all the blocktool boilerplate cli code and provide minimal support
